Launcher2: disable FadingEdge for Text Labels in Hotseat Icons
The Text Labels of App Icons in the Hotseat area are hidden to give aesthetic
effect. The label TextViews have fadingEdge effect on Marquee enabled and
eventhough the Text labels are not shown, the fading edges are created for
labels when the text is longer than the Icon width, and the fading edges are
first drawn to the buffer and then hidden. The drawing of a FadingEdge is very
costly on GPU as it might cause a resolve - leading to poor performance in the
Launcher UX.

Disabling the FadingEdge for TextViews in Icons only in the Hotseat avoids
these draws & resolves but still gives visually same functionality as the drawn
elements are hidden anyway. This change disables fadingEdge for Icons when they
are present in the Hotseat.

Change-Id: I2d2625a70490451eebc34e1e240774e47533523b