Revert changes I73d4fb85,Ie1458920 from CAF merge
Reverts:
 * commit 96360fc "Launcher: change the stype for UX launcher"
 * commit 738ab7d "Launcher: Set Apps name in main menu list marquee"

This is needed to restore the search bar. The option to remove search bar
will be added to the Launcher menu in the future.

Also commit 738ab7d causes apps with names that are just long enough to
reach the fade-out point look like the last letter is missing a part.

Change-Id: I9ad9d0c82f3a6bfde5714464dfec8662459207d8